requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.5.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pytest>=7.3.1
loguru>=0.7.0 